refactor: 인증을 위한 UserDetails 패키지 위치 변경
