+++ src/main/java/nextstep/auth/intercpetor/TokenAuthenticationInterceptor.java
@@ -8,7 +8,7 @@ import nextstep.auth.context.Authentication;
 import nextstep.auth.token.JwtTokenProvider;
 import nextstep.auth.token.TokenRequest;
 import nextstep.auth.token.TokenResponse;
-import nextstep.user.UserDetails;
+import nextstep.auth.authentication.UserDetails;
 import nextstep.user.UserDetailsService;
 import org.springframework.http.MediaType;
