+++ src/main/java/nextstep/member/application/LoginMemberService.java
@@ -3,7 +3,7 @@ package nextstep.member.application;
 import nextstep.member.domain.Member;
 import nextstep.member.domain.MemberRepository;
 import nextstep.user.User;
-import nextstep.user.UserDetails;
+import nextstep.auth.authentication.UserDetails;
 import nextstep.user.UserDetailsService;
 import org.springframework.stereotype.Service;
