+++ src/main/java/nextstep/user/UserDetailsService.java
@@ -1,5 +1,7 @@
 package nextstep.user;
 
+import nextstep.auth.authentication.UserDetails;
+
 public interface UserDetailsService {
     UserDetails loadUserByUsername(String email);
 }
